refactor BreweryQueries
